refactor(ui): move Header/Footer out of components/ into shell/
Header and Footer are site-shell compositions, not shadcn primitives —
components/ is for primitives. New import path: @syner/ui/shell/{header,footer}
(added ./shell/* to the package exports). Updated the only consumer
(apps/design home).
